requirements.txt
strands-agents>=1.0.0
bedrock-agentcore>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aws-opentelemetry-distro>=0.10.1
boto3>=1.28.0
